// src/app/features/incidents/models/incident-patch.model.ts
import { SeverityEnum } from '../enums/severity.enum';
import { StatusEnum } from '../enums/status.enum';
import { CategoryModel } from "./category.model";
import { UserRefModel } from "./user-ref.model";
import { RentalRefModel } from "./rental-ref.model";

export interface IncidentPatchModel {
  name?: string;
  reporter?: UserRefModel;
  reviewer?: UserRefModel;
  rental?: RentalRefModel;
  severity?: SeverityEnum;
  status?: StatusEnum;
  categories?: CategoryModel[];
  tags?: string;
  description?: string;
}